self-projects/farmer_assistant/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.6.0
google-adk>=1.28.0
google-generativeai>=0.5.0
googlemaps>=4.10.0
typing-extensions>=4.9.0
tzdata
pyyaml